requirements.txt
langchain
langchain_community
langchain-core
langchain-text-splitters
langchain-openai
pypdf
faiss-cpu
langchainhub
langsmith
langgraph
langchain_pinecone
requests
black
openai
google-search-results
python-dotenv
pinecone
beautifulsoup4
altair<5
streamlit
streamlit_chat
unstructured